Start handling instances, variables
This works-ish, now, and the data reaches the frontend. The problem we
get into is that we can have a data race between the creation of
instances and variables, which needs some sort of retry mechanism. One
way to do this might be to use a pool of goroutines all running the
storageUpdaterLoop (enough that we should never run out of them before
we manage to resolve a dependency loop)