Add variable retry, storage updater pool
This seems to, at a quick test, allow retries to succesfully complete
for variables while the other records are handled in the other
goroutines during the sleep. The duration etc might want to be changed,
moved elsewhere, made configurable? but for now it's there.

Similarly, the pool size might need to be configurable or at least be a
constant somewhere else.

Also, fix some incorrect comments in variable handling.