Convert Storer to interface
This serves no purpose other than allow dummying it out for storage
updater tests, right now, as the interface is identical to the interface
of the concrete implementation.

It does have the one advantage that the users of the Storer interface
don't have to concern themselves with whether the returned storer is a
pointer.